Switches to explicitly denoting attributes versus event handlers.
Before this commit, we had taken the simple route and just special-cased
a set of event handlers on the client side. However, to web components
often will use custom event names, so in order to support arbitrary web
components, we need to drive the decision of whether or not a property
is an event handler from the server.